AutoGen-Playground/requirements.txt
pyautogen~=0.2.0b5
gradio>=4.0.0
yfinance
diskcache
